Rename data -> cache

Refactor dataTransceiver classes
