+++ cpp/tensorrt_llm/batch_manager/dataTransceiver.h
@@ -21,6 +21,7 @@
 #include <map>
 #include <string>
 
+#include "tensorrt_llm/batch_manager/cacheFormatter.h"
 #include "tensorrt_llm/batch_manager/llmRequest.h"
 #include "tensorrt_llm/common/assert.h"
 #include "tensorrt_llm/common/envUtils.h"
@@ -38,6 +39,23 @@ namespace tensorrt_llm::batch_manager
 using DataContext = tensorrt_llm::executor::kv_cache::DataContext;
 using Connection = tensorrt_llm::executor::kv_cache::Connection;
 using ConnectionManager = tensorrt_llm::executor::kv_cache::ConnectionManager;
+using SizeType32 = tensorrt_llm::runtime::SizeType32;
+using TransferSession = kv_cache_manager::TransferSession;
+
+struct TransceiverTag
+{
+    enum class Id : uint64_t
+    {
+        REQUEST_SEND = 1,
+        TERMINATION = 2
+    };
+
+    static constexpr int32_t kID_TAG{19};
+    static constexpr int32_t kINFO_SIZE_TAG{22};
+    static constexpr int32_t kINFO_TAG{32};
+};
+
+using BaseCacheFormatter = kv_cache_manager::BaseCacheFormatter;
 
 // Used to store the information that needs to be sent to the context executor to ensure the generation
 // executor smoothly receives the data.
@@ -94,171 +112,19 @@ private:
     executor::DataTransceiverState mTransState;
 };
 
-class TransferSession
-{
-public:
-    struct Measure
-    {
-        double delay;     // from last token (ctx) or arrival time (gen), in ms
-        double duration;  // in ms
-        double bandwidth; // in Gbps
-    };
 
-    TransferSession(std::vector<Connection const*> connections, DataContext dataContext,
-        executor::DataTransceiverState const& selfState, executor::DataTransceiverState otherState,
-        runtime::BufferManager const& bufferManager, LlmRequest const* llmRequest = nullptr, bool recordMeasure = false)
-        : mConnections(std::move(connections))
-        , mDataContext(dataContext)
-        , mSelfState(&selfState)
-        , mOtherState(std::move(otherState))
-        , mBufferManager(&bufferManager)
-        , mRequest(llmRequest)
-        , mRecordMeasure(recordMeasure)
-    {
-        TLLM_CHECK(!mConnections.empty());
-    }
-
-    [[nodiscard]] std::vector<Connection const*> const& getConnections() const
-    {
-        return mConnections;
-    }
-
-    // should be called only during the initialization of the TransferSession
-    void setConnection(size_t idx, Connection const* conn)
-    {
-        mConnections.at(idx) = conn;
-    }
-
-    [[nodiscard]] DataContext const& getDataContext() const
-    {
-        return mDataContext;
-    }
-
-    [[nodiscard]] executor::DataTransceiverState const& getSelfState() const
-    {
-        return *mSelfState;
-    }
-
-    [[nodiscard]] executor::DataTransceiverState const& getOtherState() const
-    {
-        return mOtherState;
-    }
-
-    [[nodiscard]] runtime::BufferManager const& getBufferManager() const
-    {
-        return *mBufferManager;
-    }
-
-    void send(size_t idx, void const* data, size_t size)
-    {
-        try
-        {
-            mConnections.at(idx)->send(mDataContext, data, size);
-        }
-        catch (std::exception const& e)
-        {
-            throw common::RequestSpecificException(
-                __FILE__, __LINE__, e.what(), mRequest->mRequestId, common::RequestErrorCode::kNETWORK_ERROR);
-        }
-    }
-
-    void recv(size_t idx, void* data, size_t size)
-    {
-        try
-        {
-            mConnections.at(idx)->recv(mDataContext, data, size);
-        }
-        catch (std::exception const& e)
-        {
-            throw common::RequestSpecificException(
-                __FILE__, __LINE__, e.what(), mRequest->mRequestId, common::RequestErrorCode::kNETWORK_ERROR);
-        }
-    }
-
-    [[nodiscard]] LlmRequest const& getLlmRequest() const
-    {
-        TLLM_CHECK(mRequest != nullptr);
-        return *mRequest;
-    }
-
-    // in DataSender, the LlmRequest is not available until the sendSync is called
-    void setLlmRequest(LlmRequest const& llmRequest)
-    {
-        mRequest = &llmRequest;
-    }
-
-    void appendMeasure(double delay, double duration, size_t size);
-
-    // TODO: 1. use global id instead of context request id; 2. export to llm metrics instead of file
-    void exportMeasure(std::ofstream& outFile, bool isContext) const;
-
-private:
-    std::vector<Connection const*> mConnections;
-    DataContext mDataContext;
-    executor::DataTransceiverState const* mSelfState; // stored in DataRequester/DataResponder
-    executor::DataTransceiverState mOtherState;
-    runtime::BufferManager const* mBufferManager;
-    LlmRequest const* mRequest;
-    bool mRecordMeasure;
-    std::vector<Measure> mMeasures;
-};
-
-// Operators required for data transmission in specific communication protocols.
-class DataSender
-{
-public:
-    /// @brief Receive the request information.
-    /// @return The request information.
-    [[nodiscard]] virtual RequestInfo recvRequestInfo() = 0;
-
-    /// @brief Synchronously send data.
-    /// @param llmRequest The request object to which the data belongs.
-    virtual void sendSync(LlmRequest const& llmRequest) = 0;
-
-    /// @brief Return the internal communicator status.
-    /// @return The communicator status.
-    [[nodiscard]] virtual executor::kv_cache::CommState const& getCommState() const = 0;
-
-    /// @brief Reset the internal communicator status.
-    /// @param commState The communicator status.
-    virtual void setCommState(executor::kv_cache::CommState commState) = 0;
-
-    [[nodiscard]] virtual size_t getCounterpartsCount(LlmRequest::RequestIdType requestId) const = 0;
-
-    virtual void release(LlmRequest::RequestIdType requestId) = 0;
-
-    /// @brief Destructor.
-    virtual ~DataSender() = default;
-};
-
-// Operators required for data transmission in specific communication protocols.
-class DataReceiver
-{
-public:
-    /// @brief Send the request information.
-    /// @param llmRequest The request object to which the information belongs.
-    virtual TransferSession sendRequestInfo(LlmRequest const& llmRequest) = 0;
-
-    /// @brief Synchronously receive data.
-    /// @param session The transfer session.
-    virtual void receiveSync(TransferSession& session) = 0;
-
-    /// @brief Destructor.
-    virtual ~DataReceiver() = default;
-};
-
-class DataResponder
+class CacheSender
 {
 public:
     /// @brief Constructor.
-    /// @param sender The sender used at the underlying level.
-    explicit DataResponder(std::unique_ptr<DataSender> sender);
+    CacheSender(executor::kv_cache::ConnectionManager* manager, executor::kv_cache::CacheState selfCacheState,
+        SizeType32 selfIndex, std::unique_ptr<BaseCacheFormatter> formatter);
 
     /// @brief Asynchronously respond to the request and send data.
     /// @param llmRequest Request object. Its data should be ready when called, and the data for this request
     /// should remain valid until future synchronization.
     /// @return Once the data is fully sent, the future object will become valid.
-    [[nodiscard]] std::future<void> respondAndSendAsync(LlmRequest& llmRequest) const;
+    [[nodiscard]] std::future<void> sendAsync(LlmRequest& llmRequest) const;
 
     /// @brief Return the internal communicator status.
     /// @return The communicator status.
@@ -268,29 +134,48 @@ public:
     /// @param commState The communicator status.
     void setCommState(executor::kv_cache::CommState commState);
 
+    /// @brief Receive the request information.
+    /// @return The request information.
+    [[nodiscard]] RequestInfo recvRequestInfo();
+
+    /// @brief Synchronously send data.
+    /// @param llmRequest The request object to which the data belongs.
+    void sendSync(LlmRequest const& llmRequest);
+
+    /// @brief Send the request information.
+    /// @param llmRequest The request object to which the information belongs.
+    [[nodiscard]] TransferSession sendRequestInfo(LlmRequest const& llmRequest);
+
+    /// @brief Synchronously receive data.
+    /// @param session The transfer session.
+    void receiveSync(TransferSession& session);
+
     /// @brief Destructor.
-    ~DataResponder();
+    ~CacheSender();
 
 private:
     class Impl;
     std::unique_ptr<Impl> mImpl;
 };
 
-class DataRequester
+class CacheReceiver
 {
 public:
     /// @brief Constructor.
-    /// @param receiver The receiver used at the underlying level.
-    explicit DataRequester(std::unique_ptr<DataReceiver> receiver);
+    CacheReceiver(executor::kv_cache::ConnectionManager* manager, executor::kv_cache::CacheState selfCacheState,
+        SizeType32 selfIndex, std::unique_ptr<BaseCacheFormatter> formatter);
 
     /// @brief Asynchronously send a request to receive data.
     /// @param llmRequest Request object. Its data should be in an allocated but unwritten state when called, and the
     /// data for this request should remain intact only after future synchronization.
     /// @return Once the data is fully received, the future object will become valid.
-    [[nodiscard]] std::future<void> requestAndReceiveAsync(LlmRequest& llmRequest) const;
+    [[nodiscard]] std::future<void> receiveAsync(LlmRequest& llmRequest) const;
 
+    TransferSession sendRequestInfo(LlmRequest const& llmRequest);
+
+    void receiveSync(TransferSession& session);
     /// @brief Destructor.
-    ~DataRequester();
+    ~CacheReceiver();
 
 private:
     class Impl;
